+++ examples/Controller/BookRelationships.php
@@ -7,9 +7,10 @@ use WoohooLabs\Yin\Examples\JsonApi\Document\BookDocument;
 use WoohooLabs\Yin\Examples\JsonApi\Resource\AuthorResourceTransformer;
 use WoohooLabs\Yin\Examples\JsonApi\Resource\BookResourceTransformer;
 use WoohooLabs\Yin\Examples\JsonApi\Resource\PublisherResourceTransformer;
+use WoohooLabs\Yin\Examples\Repository\BookRepository;
 use WoohooLabs\Yin\JsonApi\Request\Request;
 
-class BookController
+class BookRelationships
 {
     /**
      * @param \Psr\Http\Message\ServerRequestInterface $request
@@ -18,30 +19,18 @@ class BookController
      */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response)
     {
-        $resource = [
-            "id" => "12345",
-            "title" => "Example Book",
-            "pages" => "200",
-            "authors" => [
-                [
-                    "id" => "11111",
-                    "name" => "John Doe"
-                ],
-                [
-                    "id" => "11112",
-                    "name" => "Jane Doe"
-                ]
-            ],
-            "publisher" => [
-                "id" => "12346",
-                "name" => "Example Publisher"
-            ]
-        ];
+        if (isset($_GET["relationship"])) {
+            $relationshipName = $_GET["relationship"];
+        } else {
+            die("You must define the 'relationship' query parameter with a value of 'authors' or 'publisher'!");
+        }
+
+        $resource = BookRepository::getBook(1);
 
         $document = new BookDocument(
             new BookResourceTransformer(new AuthorResourceTransformer(), new PublisherResourceTransformer())
         );
 
-        return $document->getResponse($response, $resource, new Request($request));
+        return $document->getRelationshipResponse($relationshipName, $response, $resource, new Request($request));
     }
 }