+++ src/JsonApi/Schema/TransformableInterface.php
@@ -1,14 +1,14 @@
 <?php
 namespace WoohooLabs\Yin\JsonApi\Schema;
 
-use WoohooLabs\Yin\JsonApi\Request\Request;
+use WoohooLabs\Yin\JsonApi\Request\RequestInterface;
 
 interface TransformableInterface
 {
     /**
      * @param mixed $resource
-     * @param \WoohooLabs\Yin\JsonApi\Request\Request $request
+     * @param \WoohooLabs\Yin\JsonApi\Request\RequestInterface $request
      * @return array
      */
-    public function transform($resource, Request $request);
+    public function transform($resource, RequestInterface $request);
 }