+++ src/JsonApi/Transformer/ResourceTransformerInterface.php
@@ -1,7 +1,7 @@
 <?php
 namespace WoohooLabs\Yin\JsonApi\Transformer;
 
-use WoohooLabs\Yin\JsonApi\Request\Request;
+use WoohooLabs\Yin\JsonApi\Request\RequestInterface;
 use WoohooLabs\Yin\JsonApi\Schema\Included;
 
 interface ResourceTransformerInterface
@@ -14,16 +14,21 @@ interface ResourceTransformerInterface
 
     /**
      * @param mixed $resource
-     * @param \WoohooLabs\Yin\JsonApi\Request\Request $request
+     * @param \WoohooLabs\Yin\JsonApi\Request\RequestInterface $request
      * @param \WoohooLabs\Yin\JsonApi\Schema\Included $included
      * @param string $baseRelationshipPath
      * @return array|null
      */
-    public function transformToResource($resource, Request $request, Included $included, $baseRelationshipPath = "");
+    public function transformToResource(
+        $resource,
+        RequestInterface $request,
+        Included $included,
+        $baseRelationshipPath = ""
+    );
 
     /**
      * @param mixed $resource
-     * @param \WoohooLabs\Yin\JsonApi\Request\Request $request
+     * @param \WoohooLabs\Yin\JsonApi\Request\RequestInterface $request
      * @param \WoohooLabs\Yin\JsonApi\Schema\Included $included
      * @param string $relationshipName
      * @param string $baseRelationshipPath
@@ -31,7 +36,7 @@ interface ResourceTransformerInterface
      */
     public function transformRelationship(
         $resource,
-        Request $request,
+        RequestInterface $request,
         Included $included,
         $relationshipName,
         $baseRelationshipPath = ""